Fix: Ignore non-existing picture for posts

+++ tierlist/posts/utils.py
@@ -20,4 +20,5 @@ def save_picture(form_picture):
 def delete_picture(picture_filename):
     picture_path = os.path.join(
         current_app.root_path, 'static/post_pics', picture_filename)
-    os.remove(picture_path)
+    if os.path.isfile(picture_path):
+        os.remove(picture_path)
